requirements.txt
# MIDOG 2021 Research - Reproducible Environment
# Python: 3.11, Platform: M1 Pro (ARM64)

# Core
numpy==1.26.4
pandas==2.2.3
scipy==1.14.1

# Image processing
pillow==10.4.0
opencv-python-headless==4.10.0.84
scikit-image==0.24.0
tifffile==2024.9.20

# Deep learning (CPU)
torch==2.5.1
torchvision==0.20.1

# Digital pathology (zarr 호환성 수정)
zarr==2.17.2
numcodecs==0.12.1
tiatoolbox==1.6.0
openslide-python==1.3.1

# ML & Stats
scikit-learn==1.5.2
statsmodels==0.14.4

# Visualization
matplotlib==3.9.2
seaborn==0.13.2

# Data augmentation
albumentations==1.4.20

# Utilities
tqdm==4.66.5
pyyaml==6.0.2
h5py==3.12.1

# Jupyter (optional)
jupyter==1.1.1
ipywidgets==8.1.5